+++ src/components/Glyph/Glyph.tsx
@@ -1,7 +1,9 @@
 import { FC, ComponentPropsWithoutRef } from "react";
 import D from "~/geometry/D";
 /**
- * Glyph can be used to render a custom
+ * Glyph can be used to render a custom SVG paths on the fly. 
+ * 
+ * I do plan on adding support to utilize more svg features directly however for now this elevates renders an svg component with a single path element. the props provided are passed to said path component
  * @param props
  * @returns 
  */
@@ -10,8 +12,24 @@ const Glyph: FC<{
 	 * The path command can be provided as a string path command or as a D path
 	 */
 	d: string | D
+
+	/**
+	 * The expected width of the glyph
+	 * 
+	 * *this property is passed to the svg props*
+	 */
 	width?: string
+
+	/**
+	 * The expected height of the glyph
+	 * 
+	 * *this property is passed to the svg props*
+	 */
 	height?: string
+
+	/**
+	 * svg props can be modified via this object.
+	 */
 	svgProps?: ComponentPropsWithoutRef<"svg">
 } & Omit<ComponentPropsWithoutRef<"path">, "d">> = ({d, width, height, svgProps:{viewBox, ...svgProps}={}, ...props}) => {
 	const p = typeof d === 'string' ? new D(d):d;
