--- src/components/types.ts
@@ -1,5 +1,0 @@
-export interface ShapleyStyles {
-	fill?: string
-	stroke?: string
-	strokeWidth?: number
-}
