+++ src/geometry/index.ts
@@ -1,5 +1,5 @@
-export { default as Command } from "./Command";
-export * from "./Command";
+export { default as Command } from "./Commands/Command";
+export * from "./Commands/Command";
 export * from "./Commands";
 export { default as D } from "./D";
 export { default as Gen } from "./Gen";
